auth_service_compose/auth_service/app/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
gunicorn==20.1.0
python-dotenv==0.20.0
pydantic==1.9.0
psycopg2-binary==2.9.3
SQLAlchemy-serializer==1.4.1
Flask-JWT-Extended==4.4.2
Werkzeug==2.1.2
Flask-WTF==1.0.1
redis==4.3.4
flask-migrate==3.1.0
flasgger==0.9.5
flask-restx==0.5.1